Code rework & minor improvements
