+++ src/families/polkadot/registry.js
@@ -5,8 +5,6 @@ import { getSpecTypes } from '@polkadot/types-known';
 import { Metadata } from '@polkadot/metadata';
 import { extrinsicsFromMeta } from '@polkadot/metadata/decorate';
 
-const POLKADOT_CHAIN_NAME = "Polkadot";
-
 // Prefix for SS58-encoded addresses on Polkadot.
 const POLKADOT_SS58_FORMAT = 0;
 
@@ -19,31 +17,24 @@ const defaultPolkadotProperties = {
     tokenSymbol: 'DOT',
 };
 
-function createMetadata(registry: TypeRegistry, metadataRpc: string): Metadata {
-    return new Metadata(registry, metadataRpc);
-}
-
-// TODO Call this only once per session
-export const createDecoratedTxs = (registry: TypeRegistry, metadataRpc: string) => {
-    return extrinsicsFromMeta(registry, createMetadata(registry, metadataRpc));
+// TODO: Put this in cache
+export const createDecoratedTxs = (registry: TypeRegistry, metadataRpc: any) => {
+    return extrinsicsFromMeta(registry, new Metadata(registry, metadataRpc));
 }
 
 /**
- * Given a a spec name, and a spec version, return the corresponding type registry for Polkadot.
- * 
+ * Given a chain name, a spec name, and a spec version, return the corresponding type registry for Polkadot.
  * @see https://github.com/polkadot-js/api/tree/master/packages/types-known
- * @param specName - The name of the runtime spec. Returned by RPC
- * `state_getRuntimeVersion`.
- * @param specVersion - The spec version of Polkadot for which we want to
- * create a type registry. Returned by RPC `state_getRuntimeVersion`.
- * @param metadataRpc - Used to run `registry.setMetadata()`
  */
-export const getRegistry = (specName: string, specVersion: string, metadataRpc: any) => {
+// TODO: Put this in cache
+export const getRegistry = (info: any) => {
+
     const registry = new TypeRegistry();
     // Register types specific to chain/runtimeVersion
-    registry.register(getSpecTypes(registry, POLKADOT_CHAIN_NAME, specName, specVersion));
+    registry.register(getSpecTypes(registry, info.chainName, info.specName, info.specVersion));
     // Register the chain properties for this registry
     registry.setChainProperties(registry.createType('ChainProperties', defaultPolkadotProperties));
-    registry.setMetadata(createMetadata(registry, metadataRpc));
+    registry.setMetadata(new Metadata(registry, info.metadataRpc));
+
     return registry;
 }
